fix: Update ig.ini for consistency
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! IMPORTANT
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
Your ig.ini file does NOT point to the Implementation Guide resource
that SUSHI generates. As a
result, the IG Publisher will IGNORE the SUSHI-generated
ImplementationGuide resource. To fix this,
please do one of the following:
- Update your ig.ini file with 'ig =
fsh-generated/resources/ImplementationGuide-example.fhir.ph.core.json'
(recommended), or
- Update your sushi-config.yaml file with 'FSHOnly: true' to tell SUSHI
NOT to generate the IG resource