Substantial refactor:
* renamed files and indentifiers in value sets and code systems to
remove the type from the name. It is not necessary.
* moved value sets and code systems into different directories:
Rationale is that valuesets are subsets designed for use, while
codesystems are (potentially external) items that should be part of a
base. An IG may create value sets for its purposes, while a server may
or may not have a code system validation installed. This is an
architectural distinction that should manifest early on.
A code system provider for $PSA for example would be implemented as a
Java class installed in a server, represented through a FHIR package on
the server.
* language is set to plain english, as en-PH is not supported in FHIR
languages, producing warnings, while es-PH is not supported by display
languages for certain code systems employed, also causing warnings.
* aliases have been created to